requirements.txt
aiofiles
alembic
asyncpg
apscheduler
boto3
dnspython
elastic_apm
fastapi
fastapi-utilities
fastapi_limiter
Flask
invoke
grpcio
langchain
langchain_aws
langchain_core
fastmcp
langchain_classic
langchain_google_genai
langchain_mcp_adapters
langchain_openai
langchain_experimental
langgraph
motor
redis
gunicorn
pymongo
pydantic
pydantic_settings
python-dotenv
python_jose
pytest
pytest-asyncio
Requests
sqlalchemy
starlette
tqdm
typesense
typing_extensions
uvicorn
websockets
valkey
arize-phoenix-otel
openinference-instrumentation-langchain
psutil
playwright
cairosvg
aiohttp
